% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGfunctions.R
\name{compare_DEGs}
\alias{compare_DEGs}
\title{Get DEG lists}
\usage{
compare_DEGs(to_compare, vst_mat, annotation, out_dir = NULL, cell = NULL)
}
\arguments{
\item{to_compare}{list of DE_df objects to compare (produced by get_DEGs)}

\item{vst_mat}{matrix of VST counts}

\item{annotation}{sample annotation df for the appropriate cell line}

\item{out_dir}{option to specify output directory}

\item{cell}{optional specification of cell line}
}
\value{
list of DE dfs
}
\description{
Get DEG lists
}
\examples{
get_DEG_lists(control_dTAG, NO_NOdTAG,...)
}
