% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGfunctions.R
\name{DEG_heatmaps}
\alias{DEG_heatmaps}
\title{DEG Heatmaps}
\usage{
DEG_heatmaps(DEG_lists, vst_mat, annotation, out_dir = NULL, cell = NULL)
}
\arguments{
\item{DEG_lists}{concatenated lits of DEG_df data}

\item{vst_mat}{matrix of VST counts}

\item{annotation}{sample annotation df for the appropriate cell line}

\item{out_dir}{option to specify output directory}

\item{cell}{option to specify cell line}
}
\value{
heatmaps
}
\description{
DEG Heatmaps
}
\examples{
DEG_heatmaps(DEG_compare_list, vst_counts)
}
