% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{get_cell_subset}
\alias{get_cell_subset}
\title{Get Cell Subset}
\usage{
get_cell_subset(cell, txi_salmon, out_dir = NULL)
}
\arguments{
\item{cell}{name of cell line to subset}

\item{txi_salmon}{txi list}

\item{out_dir}{option to specify output directory}
}
\value{
a list containing two objects: an annotations df and corresponding matrix of abundances
}
\description{
Get Cell Subset
}
\examples{
get_cell_subset('Kasumi', txi_list)
}
