% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGfunctions.R
\name{get_condition_DEGs}
\alias{get_condition_DEGs}
\title{Get DEGs for one Condition}
\usage{
get_condition_DEGs(
  dds,
  c1,
  c2,
  organism = "hsapiens",
  cell = NULL,
  out_dir = NULL
)
}
\arguments{
\item{dds}{DESeqDataSet object}

\item{c1}{condition 1, e.g. 'control'}

\item{c2}{condition 2, e.g. 'dTAG'}

\item{organism}{e.g. 'hsapiens', for conversion of ENSG IDs to gene symbols}

\item{cell}{option to specify cell line for output naming}

\item{out_dir}{option to specify output directory}
}
\value{
list of 2, DE matrix and table of DEGs
}
\description{
Get DEGs for one Condition
}
\examples{
get_condition_DEGs(kasumi_dds_object, 'control', 'dTAG', cell='Kasumi' )
}
