% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{get_ref}
\alias{get_ref}
\title{Get Referemce}
\usage{
get_ref(gtf_dir)
}
\arguments{
\item{gtf_dir}{Directory where GTF file is located}
}
\value{
A txdb object that stores transcript metadata
}
\description{
Get Referemce
}
\examples{
get_ref('/path/to/gtf_file')
}
