% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{make_annotation}
\alias{make_annotation}
\title{Make Annotation, import_salmon helper function}
\usage{
make_annotation(salmon_dir)
}
\arguments{
\item{salmon_dir}{Directory where salmon output file folders are located}
}
\value{
A data frame of sample annotations
}
\description{
Make Annotation, import_salmon helper function
}
\examples{
make_annotation('/path/to/salmon_folders')
}
