% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCpipeline.R
\name{QC_DESeq}
\alias{QC_DESeq}
\title{Quality control and DESeq}
\usage{
QC_DESeq(
  gtf_dir,
  salmon_dir,
  out_dir = NULL,
  cell = NULL,
  remove_batch = TRUE,
  grp_by = "condition"
)
}
\arguments{
\item{gtf_dir}{Directory where salmon output file folders are located}

\item{salmon_dir}{Data frame containing sample annotations}

\item{out_dir}{option to specify directory to which figures and tables will be exported}

\item{cell}{optional name of cell line to subset}

\item{remove_batch}{option to remove batch effects (assume replicate=batch)}

\item{grp_by}{attribute to group by, e.g. 'condition', 'replicate', etc.}
}
\value{
list of DESeqDataSet, VST DESeqDataSet, VST counts, and annotation df
}
\description{
Quality control and DESeq
}
\examples{
QC_DESeq('~/refDirectory', '~/salmonDirectory', cell='Kasumi')
}
