% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeqFunctions.R
\name{run_DESeq,list-method}
\alias{run_DESeq,list-method}
\title{DESeq with list input}
\usage{
\S4method{run_DESeq}{list}(
  salmon_data,
  annotation,
  cell = NULL,
  batch_effect = TRUE,
  out_dir = NULL
)
}
\arguments{
\item{salmon_data}{salmon tximport list}

\item{annotation}{sample annotation df}

\item{cell}{option to specify cell line}

\item{batch_effect}{option to remove batch effects (assume replicate=batch)}

\item{out_dir}{option to specify output directory}
}
\value{
list containing DESeqDataSet object, VST version of theDESeqDataSet, and VST counts
}
\description{
DESeq with list input
}
\examples{
run_DESeq(txi_list, annotation_df)
}
