% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonQC.R
\name{salmon_QC,list-method}
\alias{salmon_QC,list-method}
\title{Salmon Quality Control from txi list}
\usage{
\S4method{salmon_QC}{list}(
  salmon_data,
  dds,
  vst,
  annotation,
  grp_by = "condition",
  out_dir = NULL,
  cell = NULL
)
}
\arguments{
\item{salmon_data}{txi list of salmon data}

\item{dds}{DESeqDataSet object}

\item{vst}{VST DESeqDataSet}

\item{annotation}{sample annotation df}

\item{grp_by}{attribute to group by, e.g. 'condition', 'replicate', etc.}

\item{out_dir}{option to specify directory to which figures and tables will be exported}

\item{cell}{option to specify cell line}
}
\value{
QC figures and tables
}
\description{
Salmon Quality Control from txi list
}
\examples{
salmon_QC(txi_list, dds, vst, annotation_df)
}
