% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{txi_salmon}
\alias{txi_salmon}
\title{txi_salmon, import_salmon helper function}
\usage{
txi_salmon(salmon_dir, annotation, tx2gene)
}
\arguments{
\item{salmon_dir}{Directory where salmon output file folders are located}

\item{annotation}{Data frame containing sample annotations}

\item{tx2gene}{A txdb object}
}
\value{
a (txi) list containing matrices extracted from salmon output: abundance, counts, length
}
\description{
txi_salmon, import_salmon helper function
}
\examples{
txi_salmon('~/salmonDirectory',annotation_dir,tx2gene)
}
